fix delete() and add getQueue() methods
delete() must call parent delete() in order to properly mark the job deleted. getQueue() must be implemented since the inherited $queue which should be a string is being replaced by the AMQPQueue object. this should actually be changed to its own $amqp variable and $queue should always be a string